Move the various character type modifiers to their own file.
