+++ bot.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 import inspect
 import itertools
@@ -130,7 +132,7 @@ class Game:
         )
         await announcement.pin()
 
-    async def start_day(self, kills=None, origin=None):
+    async def start_day(self, kills: list[Player] = None, origin=None):
         if kills is None:
             kills = []
 
@@ -449,8 +451,8 @@ class Vote:
     # Stores information about a specific vote
 
     def __init__(self, nominee, nominator):
-        self.nominee = nominee
-        self.nominator = nominator
+        self.nominee: Player = nominee
+        self.nominator: Player = nominator
         if self.nominee is not None:
             ordered_voters = (
                 game.seatingOrder[game.seatingOrder.index(self.nominee) + 1:]
@@ -468,13 +470,13 @@ class Vote:
 
         self.order = order_with_banshees
 
-        self.votes = 0
-        self.voted = []
+        self.votes: int = 0
+        self.voted: list[Player] = []
         self.history = []
         self.announcements = []
         self.presetVotes = {}
-        self.values = {person: (0, 1) for person in self.order}
-        self.majority = 0.0
+        self.values: dict[Player, tuple[int, int]] = {person: (0, 1) for person in self.order}
+        self.majority: float = 0.0
         for person in self.order:
             if not person.isGhost:
                 self.majority += 0.5
@@ -943,7 +945,8 @@ class Player:
 
     async def kill(self, suppress=False, force=False):
         dies = True
-        on_death_characters = sorted([person.character for person in game.seatingOrder if isinstance(person.character, DeathModifier)], key=lambda c: c.on_death_priority())
+        on_death_characters = sorted([person.character for person in game.seatingOrder if isinstance(person.character,
+                                                                                                     DeathModifier)], key=lambda c: c.on_death_priority())
         for player_character in on_death_characters:
             dies = player_character.on_death(self, dies)
 
@@ -1201,252 +1204,14 @@ class Player:
             pass
 
 
-class SeatingOrderModifier(Character):
-    # A character which modifies the seating order or seating order message
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    def seating_order(self, seatingOrder):
-        # returns a seating order after the character's modifications
-        return seatingOrder
-
-    def seating_order_message(self, seatingOrder):
-        # returns a string to be added to the seating order message specifically (not just to the seating order)
-        return ""
-
-
-class DayStartModifier(Character):
-    # A character which modifies the start of the day
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    async def on_day_start(self, origin, kills):
-        # Called on the start of the day
-        return True
-
-
-class NomsCalledModifier(Character):
-    # A character which modifies the start of the day
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    def on_noms_called(self):
-        # Called when nominations are called for the first time each day
-        pass
-
-
-class NominationModifier(Character):
-    # A character which triggers on a nomination
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    async def on_nomination(self, nominee, nominator, proceed):
-        # Returns bool -- whether the nomination proceeds
-        return proceed
-
-
-class DayEndModifier(Character):
-    # A character which modifies the start of the day
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    def on_day_end(self):
-        # Called on the end of the day
-        pass
-
-
-class VoteBeginningModifier(Character):
-    # A character which modifies the value of players' votes
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    def modify_vote_values(self, order, values, majority):
-        # returns a list of the vote's order, a dictionary of vote values, and majority
-        return order, values, majority
-
-
-class VoteModifier(Character):
-    # A character which modifies the effect of votes
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    def on_vote_call(self, toCall):
-        # Called every time a player is called to vote
-        pass
-
-    def on_vote(self):
-        # Called every time a player votes
-        pass
-
-    def on_vote_conclusion(self, dies, tie):
-        # returns boolean -- whether the nominee is about to die, whether the vote is tied
-        return dies, tie
-
-
-class DeathModifier(Character):
-    # A character which triggers on a player's death
-    PROTECTS_OTHERS = 1
-    PROTECTS_SELF = 2
-    KILLS_SELF = 3
-    FORCES_KILL = 1000
-    UNSET = 999
-
-    def __init__(self, parent):
-        super().__init__(parent)
-
-    def on_death(self, person, dies):
-        # Returns bool -- does person die
-        return dies
-
-    def on_death_priority(self):
-        return DeathModifier.UNSET
-
-
-class AbilityModifier(
-    SeatingOrderModifier,
-    DayStartModifier,
-    NomsCalledModifier,
-    NominationModifier,
-    DayEndModifier,
-    VoteBeginningModifier,
-    VoteModifier,
-    DeathModifier,
-):
-    # A character which can have different abilities
-
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.abilities = []
-
-    def refresh(self):
-        super().refresh()
-        self.abilities = []
-
-    def add_ability(self, role):
-        self.abilities.append(role(self.parent))
-
-    def clear_ability(self):
-        removed_ability = None
-        for ability in self.abilities:
-            if isinstance(ability, AbilityModifier):
-                removed_ability = ability.clear_ability()
-        if not removed_ability:
-            if len(self.abilities):
-                removed_ability = self.abilities.pop()
-        return removed_ability
-
-    def seating_order(self, seatingOrder):
-        # returns a seating order after the character's modifications
-        for role in self.abilities:
-            if isinstance(role, SeatingOrderModifier):
-                seatingOrder = role.seating_order(seatingOrder)
-        return seatingOrder
-
-    async def on_day_start(self, origin, kills):
-        # Called on the start of the day
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, DayStartModifier):
-                    await role.on_day_start(origin, kills)
-
-        return True
-
-    def poison(self):
-        super().poison()
-        for role in self.abilities:
-            role.poison()
-
-    def unpoison(self):
-        super().unpoison()
-        for role in self.abilities:
-            role.unpoison()
-
-    def on_noms_called(self):
-        # Called when nominations are called for the first time each day
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, NomsCalledModifier):
-                    role.on_noms_called()
-
-    async def on_nomination(self, nominee, nominator, proceed):
-        # Returns bool -- whether the nomination proceeds
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, NominationModifier):
-                    proceed = await role.on_nomination(nominee, nominator, proceed)
-        return proceed
-
-    def on_day_end(self):
-        # Called on the end of the day
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, DayEndModifier):
-                    role.on_day_end()
-
-    def modify_vote_values(self, order, values, majority):
-        # returns a list of the vote's order, a dictionary of vote values, and majority
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, VoteBeginningModifier):
-                    order, values, majority = role.modify_vote_values(
-                        order, values, majority
-                    )
-        return order, values, majority
-
-    def on_vote_call(self, toCall):
-        # Called every time a player is called to vote
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, VoteModifier):
-                    role.on_vote_call(toCall)
-
-    def on_vote(self):
-        # Called every time a player votes
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, VoteModifier):
-                    role.on_vote()
-
-    def on_vote_conclusion(self, dies, tie):
-        # returns boolean -- whether the nominee is about to die, whether the vote is tied
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, VoteModifier):
-                    dies, tie = role.on_vote_conclusion(dies, tie)
-        return dies, tie
-
-    def on_death(self, person, dies):
-        # Returns bool -- does person die
-        if not self.is_poisoned:
-            for role in self.abilities:
-                if isinstance(role, DeathModifier):
-                    dies = role.on_death(person, dies)
-        return dies
-
-    def on_death_priority(self):
-        priority = DeathModifier.UNSET
-        if not self.is_poisoned and not self.parent.isGhost:
-            for role in self.abilities:
-                if isinstance(role, DeathModifier):
-                    priority = min(priority, role.on_death_priority())
-        return priority
-
-
 class Traveler(SeatingOrderModifier):
     # A generic traveler
 
-    def __init__(self, parent):
+    def __init__(self, parent: Player):
         super().__init__(parent)
         self.role_name = "Traveler"
 
-    def seating_order_message(self, seatingOrder):
+    def seating_order_message(self, seating_order):
         return " - {}".format(self.role_name)
 
     async def exile(self, person, user):
@@ -1523,7 +1288,7 @@ class Storyteller(SeatingOrderModifier):
         super().__init__(parent)
         self.role_name = "Storyteller"
 
-    def seating_order_message(self, seatingOrder):
+    def seating_order_message(self, seating_order):
         return " - {}".format(self.role_name)
