+++ model/characters/__init__.py
@@ -1,4 +1,8 @@
 from .character import Character
 from .character_types import *
+from .character_modifiers import *
 
-__all__ = ["Character", "Townsfolk", "Outsider", "Minion", "Demon"]
+__all__ = ["Character",
+           "Townsfolk", "Outsider", "Minion", "Demon",
+           "AbilityModifier", "DayEndModifier", "DayStartModifier", "DeathModifier", "NomsCalledModifier",
+           "NominationModifier", "SeatingOrderModifier", "VoteBeginningModifier", "VoteModifier"]
