+++ model/characters/character.py
@@ -1,10 +1,15 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from bot import Player
+
 
 class Character:
     # A generic character
-    def __init__(self, parent: 'Player'):
-        self.parent: 'Player' = parent
+    def __init__(self, parent: Player):
+        self.parent: Player = parent
         self.role_name: str = "Character"
         self._is_poisoned: bool = False
         self.refresh()
